requirements.txt
python-telegram-bot[all]==20.7
geopy==2.4.0
python-dotenv==1.0.1
